refactor(#97): 사용자 닉네임 조회 책임을 auth 서비스로 분리
